refactor: prepare wasm target
