+++ package/origlang-cli/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "origlang-cli"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = { version = "4.2.1", features = ["derive"] }
+env_logger = "0.10.0"
+log = "0.4.17"
+thiserror = "1.0.39"
+origlang-compiler = { path = "../origlang-compiler" }
+origlang-typesystem-model = { path = "../origlang-typesystem-model" }
+origlang-runtime = { path = "../origlang-runtime" }
+ariadne = "0.2.0"
+origlang-platform = { path = "../origlang-platform" }
