+++ package/origlang-compiler/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "origlang-compiler"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+origlang-ast = { path = "../origlang-ast" }
+origlang-typesystem-model = { path = "../origlang-typesystem-model" }
+derive_more = "0.99.17"
+log = "0.4.17"
+thiserror = "1.0.39"
+origlang-platform = { path = "../origlang-platform" }
