+++ package/origlang-compiler/src/error.rs
@@ -1,0 +1,11 @@
+use thiserror::Error;
+use crate::parser::SimpleErrorWithPos;
+use crate::type_check::error::TypeCheckError;
+
+#[derive(Error, Debug)]
+pub enum AllError {
+    #[error("{0}")]
+    Generic(#[from] SimpleErrorWithPos),
+    #[error("{0}")]
+    TypeCheck(#[from] TypeCheckError),
+}
