+++ package/origlang-compiler/src/lib.rs
@@ -1,0 +1,5 @@
+mod char_list;
+mod error;
+pub mod lexer;
+pub mod parser;
+pub mod type_check;
