--- package/origlang-core/src/cli.rs
@@ -1,2 +1,0 @@
-pub mod args;
-mod task;
