--- package/origlang-core/src/cli/task/interpret.rs
@@ -1,24 +1,0 @@
-use crate::cli::task::print_ast::ParseSource;
-use crate::cli::task::Task;
-use crate::error::AllError;
-use crate::runtime::{PrintToStdout, Runtime};
-use crate::type_check::TypeChecker;
-
-pub struct Interpret;
-
-impl Task for Interpret {
-    type Environment = ParseSource;
-    type Error = AllError;
-
-    fn execute(&self, environment: Self::Environment) -> Result<(), Self::Error> {
-        use crate::parser::Parser;
-        let source = environment.source();
-        let parser = Parser::create(source.as_str());
-        let root_ast = parser.parse()?;
-        let type_checker = TypeChecker::new();
-        type_checker.check(&root_ast)?;
-        let runtime = Runtime::create(PrintToStdout);
-        runtime.execute(root_ast);
-        Ok(())
-    }
-}
