+++ package/origlang-interop/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "origlang-interop"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+wasm-bindgen = "0.2.84"
