+++ package/origlang-platform/src/lib.rs
@@ -1,0 +1,7 @@
+mod platform;
+
+#[cfg(windows)]
+pub use platform::windows::*;
+
+#[cfg(not(windows))]
+pub use platform::non_windows::*;
