+++ package/origlang-platform/src/platform.rs
@@ -1,0 +1,5 @@
+#[cfg(windows)]
+pub mod windows;
+
+#[cfg(not(windows))]
+pub mod non_windows;
