+++ package/origlang-runtime/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "origlang-runtime"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+thiserror = "1.0.39"
+derivative = "2.2.0"
+derive_more = "0.99.17"
+log = "0.4.17"
+tap = "1.0.1"
+origlang-typesystem-model = { path = "../origlang-typesystem-model" }
+origlang-ast = { path = "../origlang-ast" }
